requirements.txt
# Core Dependencies
supabase>=1.0.0
sentence-transformers>=2.2.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Document Processing Dependencies
python-docx>=0.8.11          # Word documents (.docx)
openpyxl>=3.1.0              # Excel files (.xlsx)
pdfplumber>=0.9.0            # PDF processing (primary)
PyPDF2>=3.0.0               # PDF processing (fallback)
beautifulsoup4>=4.12.0       # HTML parsing
python-pptx>=0.6.21          # PowerPoint files (.pptx)
ebooklib>=0.18               # EPUB files

# Optional Dependencies for Enhanced Features
lxml>=4.9.0                  # XML processing (beautifulsoup backend)
Pillow>=10.0.0               # Image processing
python-magic>=0.4.27        # File type detection
chardet>=5.0.0               # Character encoding detection

# Development Dependencies (optional)
pytest>=7.0.0               # Testing framework
black>=23.0.0               # Code formatting
flake8>=6.0.0               # Code linting
mypy>=1.0.0                 # Type checking

# Additional ML/AI Dependencies
scikit-learn>=1.3.0         # Machine learning utilities
transformers>=4.30.0        # Advanced transformer models (optional)
torch>=2.0.0                # PyTorch for advanced models (optional)